Move the utility functions to utils for now.
