+++ web/app/utils/utils.js
@@ -1,3 +1,4 @@
+import {createAction as createReduxAction} from 'redux-actions'
 
 /**
  * Wraps an action creator function so that the React synthetic action
@@ -60,3 +61,18 @@ export const requestIdleCallback = (fn) => {
         return setTimeout(() => fn(), 1)
     }
 }
+
+const typecheck = (type, value) => {
+    try {
+        type.check(value)
+    } catch (e) {
+        console.info(e)
+        console.log(value)
+    }
+    return value
+}
+
+export const createTypedAction = (description, type) => createReduxAction(
+    description,
+    (payload) => typecheck(type, payload)
+)
